// src/app/shared/enums/izvestaji.ts
export enum Izvestaji {
  BILANS_STANJA = 'Bilans Stanja',
  BILANS_USPEHA = 'Bilans Uspeha',
  PROMENE_KAPITALA ='Izveštaj o promenama na kapitalu',
  STATICKI_IZVESTAJ_KOMITET = 'Izveštaj o transakcijama vezanim za komitenta',
  STATICKI_IZVESTAJ_PO_SIFRI = 'Statistički izveštaj o transakcijama po šifri transakcije'
}